# [Opt] Optimize hotspot functions in compaction