--- a/be/src/olap/tablet_meta.h
+++ b/be/src/olap/tablet_meta.h
@@ -123,6 +123,7 @@ public:
     // disk space occupied by tablet
     size_t tablet_footprint() const;
     size_t version_count() const;
+    void update_max_version();
     Version max_version() const;
 
     TabletState tablet_state() const;
@@ -198,7 +199,8 @@ private:
     // the reference of _schema may use in tablet, so here need keep
     // the lifetime of tablemeta and _schema is same with tablet
     std::shared_ptr<TabletSchema> _schema;
-
+    // _maxVersion record _rs_metas's max endVersion Version
+    Version _maxVersion = {-1, 0};
     std::vector<RowsetMetaSharedPtr> _rs_metas;
     // This variable _stale_rs_metas is used to record these rowsets‘ meta which are be compacted.
     // These stale rowsets meta are been removed when rowsets' pathVersion is expired,