@zbtzbtzbt 2022-04-08
I also considered using SIMD max, 
but O(1) is fastest, and size of vector is not 128X or 256X.
so this way is better.